# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProviEmplea API",
        "description": "API REST para la plataforma de empleo ProviEmplea de Providencia. Permite gestionar talentos, empresas y procesos de selección con curriculum ciego.\n\n## Rate Limiting\nLa API implementa limitación de solicitudes para garantizar disponibilidad:\n- **Endpoints públicos:** 60 solicitudes por minuto\n- **Endpoints administrativos:** 30 solicitudes por minuto\n\n## Caché\nAlgunos endpoints utilizan caché para mejorar el rendimiento:\n- **GET /personas:** caché de 5 minutos\n- **GET /empresas:** caché de 5 minutos\n- **GET /admin/estadisticas:** caché de 10 minutos\n\n## Tiempos de respuesta esperados\n- Operaciones de lectura: < 200ms\n- Operaciones de escritura: < 500ms",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/empresas": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Listar empresas validadas",
                "operationId": "getEmpresas",
                "parameters": [
                    {
                        "name": "tipo_empresa",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "contratacion-directa",
                                "est",
                                "outsourcing"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Empresa"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Registrar nueva empresa",
                "operationId": "createEmpresa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpresaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/empresas/{id}": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Obtener empresa por ID",
                "operationId": "getEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Empresa"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Actualizar empresa",
                "operationId": "updateEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmpresaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empresa actualizada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Desactivar empresa",
                "description": "Desactiva el perfil sin eliminarlo de la base de datos.",
                "operationId": "deleteEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa desactivada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        },
        "/empresas/{id}/validar": {
            "patch": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Validar empresa (solo administración)",
                "operationId": "validarEmpresa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa validada"
                    },
                    "404": {
                        "description": "No encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmpresaInput": {
                "title": "Empresa Input",
                "description": "Datos para crear o actualizar una empresa",
                "properties": {
                    "nombre_empresa": {
                        "type": "string",
                        "example": "TechCorp SpA"
                    },
                    "rut_empresa": {
                        "type": "string",
                        "example": "76123456-7"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "rrhh@techcorp.cl"
                    },
                    "logo_url": {
                        "type": "string",
                        "example": "https://techcorp.cl/logo.png",
                        "nullable": true
                    },
                    "rubro": {
                        "type": "string",
                        "example": "Tecnología",
                        "nullable": true
                    },
                    "tipo_empresa": {
                        "type": "string",
                        "enum": [
                            "contratacion-directa",
                            "est",
                            "outsourcing"
                        ],
                        "example": "contratacion-directa"
                    },
                    "presentacion": {
                        "type": "string",
                        "example": "Empresa de desarrollo de software con 10 años de experiencia",
                        "nullable": true
                    },
                    "beneficios": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Seguro complementario",
                            "Trabajo remoto"
                        ],
                        "nullable": true
                    },
                    "contacto_nombre": {
                        "type": "string",
                        "example": "Ana López"
                    },
                    "contacto_email": {
                        "type": "string",
                        "format": "email",
                        "example": "ana@techcorp.cl"
                    },
                    "contacto_telefono": {
                        "type": "string",
                        "example": "+56912345678",
                        "nullable": true
                    }
                },
                "type": "object",
                "example": {
                    "nombre_empresa": "TechCorp SpA",
                    "rut_empresa": "76123456-7",
                    "email": "rrhh@techcorp.cl",
                    "rubro": "Tecnología",
                    "tipo_empresa": "contratacion-directa",
                    "presentacion": "Empresa de desarrollo de software con 10 años de experiencia",
                    "beneficios": [
                        "Seguro complementario",
                        "Trabajo remoto"
                    ],
                    "contacto_nombre": "Ana López",
                    "contacto_email": "ana@techcorp.cl",
                    "contacto_telefono": "+56912345678"
                }
            },
            "Empresa": {
                "title": "Empresa",
                "description": "Modelo completo de empresa empleadora",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "nombre_empresa": {
                        "type": "string",
                        "example": "TechCorp SpA"
                    },
                    "rut_empresa": {
                        "type": "string",
                        "example": "76123456-7"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "rrhh@techcorp.cl"
                    },
                    "logo_url": {
                        "type": "string",
                        "example": "https://techcorp.cl/logo.png",
                        "nullable": true
                    },
                    "rubro": {
                        "type": "string",
                        "example": "Tecnología",
                        "nullable": true
                    },
                    "tipo_empresa": {
                        "type": "string",
                        "enum": [
                            "contratacion-directa",
                            "est",
                            "outsourcing"
                        ],
                        "example": "contratacion-directa"
                    },
                    "presentacion": {
                        "type": "string",
                        "example": "Empresa de desarrollo de software con 10 años de experiencia",
                        "nullable": true
                    },
                    "beneficios": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Seguro complementario",
                            "Trabajo remoto",
                            "Bono de desempeño"
                        ],
                        "nullable": true
                    },
                    "contacto_nombre": {
                        "type": "string",
                        "example": "Ana López"
                    },
                    "contacto_email": {
                        "type": "string",
                        "format": "email",
                        "example": "ana@techcorp.cl"
                    },
                    "contacto_telefono": {
                        "type": "string",
                        "example": "+56912345678",
                        "nullable": true
                    },
                    "validado": {
                        "type": "boolean",
                        "example": true
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-01-10T09:00:00+00:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-05-20T08:15:00+00:00"
                    }
                },
                "type": "object",
                "example": {
                    "id": "550e8400-e29b-41d4-a716-446655440001",
                    "nombre_empresa": "TechCorp SpA",
                    "rut_empresa": "76123456-7",
                    "email": "rrhh@techcorp.cl",
                    "rubro": "Tecnología",
                    "tipo_empresa": "contratacion-directa",
                    "presentacion": "Empresa de desarrollo de software con 10 años de experiencia",
                    "beneficios": [
                        "Seguro complementario",
                        "Trabajo remoto",
                        "Bono de desempeño"
                    ],
                    "contacto_nombre": "Ana López",
                    "contacto_email": "ana@techcorp.cl",
                    "contacto_telefono": "+56912345678",
                    "validado": true,
                    "activo": true
                }
            },
...
        }
    },
...
}